refactor(core): reorg and separate logto-config APIs into files
