+++ packages/core/src/libraries/sign-in-experience/index.test.ts
@@ -59,6 +59,7 @@ const cloudConnection = createCloudConnectionLibrary({
   getOidcConfigs: jest.fn(),
   upsertJwtCustomizer: jest.fn(),
   getJwtCustomizer: jest.fn(),
+  getJwtCustomizers: jest.fn(),
 });
 
 const getLogtoConnectors = jest.spyOn(connectorLibrary, 'getLogtoConnectors');
