+++ packages/core/src/routes/init.ts
@@ -25,7 +25,7 @@ import domainRoutes from './domain.js';
 import hookRoutes from './hook.js';
 import interactionRoutes from './interaction/index.js';
 import logRoutes from './log.js';
-import logtoConfigRoutes from './logto-config.js';
+import logtoConfigRoutes from './logto-config/index.js';
 import organizationRoutes from './organization/index.js';
 import resourceRoutes from './resource.js';
 import resourceScopeRoutes from './resource.scope.js';
