+++ packages/core/src/routes/logto-config/index.test.ts
@@ -1,15 +1,9 @@
-import { LogtoOidcConfigKey, type AdminConsoleData, LogtoJwtTokenKey } from '@logto/schemas';
+import { LogtoOidcConfigKey, type AdminConsoleData } from '@logto/schemas';
 import { generateStandardId } from '@logto/shared';
 import { createMockUtils, pickDefault } from '@logto/shared/esm';
 import Sinon from 'sinon';
 
-import {
-  mockAdminConsoleData,
-  mockCookieKeys,
-  mockPrivateKeys,
-  mockLogtoConfigRows,
-  mockJwtCustomizerConfigForAccessToken,
-} from '#src/__mocks__/index.js';
+import { mockAdminConsoleData, mockCookieKeys, mockPrivateKeys } from '#src/__mocks__/index.js';
 import { MockTenant } from '#src/test-utils/tenant.js';
 import { createRequester } from '#src/utils/test-utils.js';
 
@@ -53,8 +47,6 @@ const logtoConfigQueries = {
     },
   }),
   updateOidcConfigsByKey: jest.fn(),
-  getRowsByKeys: jest.fn(async () => mockLogtoConfigRows),
-  deleteJwtCustomizer: jest.fn(),
 };
 
 const logtoConfigLibraries = {
@@ -62,11 +54,9 @@ const logtoConfigLibraries = {
     [LogtoOidcConfigKey.PrivateKeys]: mockPrivateKeys,
     [LogtoOidcConfigKey.CookieKeys]: mockCookieKeys,
   })),
-  upsertJwtCustomizer: jest.fn(),
-  getJwtCustomizer: jest.fn(),
 };
 
-const settingRoutes = await pickDefault(import('./logto-config.js'));
+const settingRoutes = await pickDefault(import('./index.js'));
 
 describe('configs routes', () => {
   const tenantContext = new MockTenant(undefined, { logtoConfigs: logtoConfigQueries });
@@ -226,62 +216,5 @@ describe('configs routes', () => {
       LogtoOidcConfigKey.PrivateKeys,
       [newPrivateKey2, newPrivateKey]
     );
-  });
-
-  it('PUT /configs/jwt-customizer/:tokenType should add a record successfully', async () => {
-    logtoConfigQueries.getRowsByKeys.mockResolvedValueOnce({
-      ...mockLogtoConfigRows,
-      rows: [],
-      rowCount: 0,
-    });
-    logtoConfigLibraries.upsertJwtCustomizer.mockResolvedValueOnce(
-      mockJwtCustomizerConfigForAccessToken
-    );
-    const response = await routeRequester
-      .put(`/configs/jwt-customizer/access-token`)
-      .send(mockJwtCustomizerConfigForAccessToken.value);
-    expect(logtoConfigLibraries.upsertJwtCustomizer).toHaveBeenCalledWith(
-      LogtoJwtTokenKey.AccessToken,
-      mockJwtCustomizerConfigForAccessToken.value
-    );
-    expect(response.status).toEqual(201);
-    expect(response.body).toEqual(mockJwtCustomizerConfigForAccessToken.value);
-  });
-
-  it('PUT /configs/jwt-customizer/:tokenType should update a record successfully', async () => {
-    logtoConfigQueries.getRowsByKeys.mockResolvedValueOnce({
-      ...mockLogtoConfigRows,
-      rows: [mockJwtCustomizerConfigForAccessToken],
-      rowCount: 1,
-    });
-    logtoConfigLibraries.upsertJwtCustomizer.mockResolvedValueOnce(
-      mockJwtCustomizerConfigForAccessToken
-    );
-    const response = await routeRequester
-      .put('/configs/jwt-customizer/access-token')
-      .send(mockJwtCustomizerConfigForAccessToken.value);
-    expect(logtoConfigLibraries.upsertJwtCustomizer).toHaveBeenCalledWith(
-      LogtoJwtTokenKey.AccessToken,
-      mockJwtCustomizerConfigForAccessToken.value
-    );
-    expect(response.status).toEqual(200);
-    expect(response.body).toEqual(mockJwtCustomizerConfigForAccessToken.value);
-  });
-
-  it('GET /configs/jwt-customizer/:tokenType should return the record', async () => {
-    logtoConfigLibraries.getJwtCustomizer.mockResolvedValueOnce(
-      mockJwtCustomizerConfigForAccessToken.value
-    );
-    const response = await routeRequester.get('/configs/jwt-customizer/access-token');
-    expect(response.status).toEqual(200);
-    expect(response.body).toEqual(mockJwtCustomizerConfigForAccessToken.value);
-  });
-
-  it('DELETE /configs/jwt-customizer/:tokenType should delete the record', async () => {
-    const response = await routeRequester.delete('/configs/jwt-customizer/client-credentials');
-    expect(logtoConfigQueries.deleteJwtCustomizer).toHaveBeenCalledWith(
-      LogtoJwtTokenKey.ClientCredentials
-    );
-    expect(response.status).toEqual(204);
   });
 });
