+++ packages/integration-tests/src/__mocks__/jwt-customizer.ts
@@ -1,0 +1,15 @@
+export const clientCredentialsJwtCustomizerPayload = {
+  script: '',
+  envVars: {},
+  contextSample: {},
+};
+
+export const accessTokenJwtCustomizerPayload = {
+  ...clientCredentialsJwtCustomizerPayload,
+  contextSample: {
+    user: {
+      username: 'test',
+      id: 'fake-id',
+    },
+  },
+};
